pyproject.toml
[tool.poetry]
name = "testing-agents"
version = "0.1.0"
description = "ADK Agents - Multiple agent implementations"
authors = ["dxscale"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
google-adk = "^1.24.1"
deprecated = "^1.2"
litellm = "^1.72.4"

[tool.poetry.group.dev.dependencies]
ipykernel = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
